Fix: Clean cmd Exception when using MongoDB Docs
findAndRemove() find the first element and return it
This leads to an exception when trying to count it in L54 of the CleanCommand.php

Add: Return the correct number of deleted rows

Thanks to @jmikola,
We're using the Query Option "safe" to get the number of rows modified by the query.

When executing the query, it now returns the following Array:
Array
(
    [n] => 0
    [connectionId] => 113959
    [err] =>
    [ok] => 1
)

Where n is the number of deleted rows.