Allow to use an external fos_oauth_server.storage service
The fact that the definition of the fos_oauth_server.storage service was
asked in the extension object forbade to use a service defined in an
other bundle. Indeed, all the definitions of the container are not
merged in the extension object and the scope of the definitions in this
file is only the bundle itself.